+++ tests/testthat/test-ifcb_psd.R
@@ -85,7 +85,8 @@ test_that("ifcb_psd works correctly", {
     missing_cells = 0.7,
     biomass = 1000,
     bloom = 5,
-    humidity = 99
+    humidity = 99,
+    use_plot_subfolders = TRUE
   )
 
   # Verify that the output list contains data, fits, and flags